Fix multi-policy related issues.
The "rule" and "expected_error_code" parameters of the rbac_rule_validation
decorator have been removed after a deprecation period, and this means
that any test that uses "rule" or "expected_error_code" parameters
will need to be changed to use "rules" and "expected_error_codes".

This is as par Patrole commit
https://review.openstack.org/#/c/595450/

Also, fixed tox pep8 issue as well as with a new .gitignore file.